Fix issue 13+16 (#17)
* address bug in TypeSchemaMapper

* address key identifier data type range issue

* version bump to 0.4.2 patch release

fixes #13 
fixes #16

+++ connect-transform-kryptonite/src/main/java/com/github/hpgrahsl/kafka/connect/transforms/kryptonite/TypeSchemaMapper.java
@@ -46,7 +46,7 @@ public interface TypeSchemaMapper {
           Type.INT8, SchemaBuilder::int8,
           Type.INT16, SchemaBuilder::int16,
           Type.INT32, SchemaBuilder::int32,
-          Type.INT64, SchemaBuilder::int16,
+          Type.INT64, SchemaBuilder::int64,
           Type.FLOAT32, SchemaBuilder::float32,
           Type.FLOAT64, SchemaBuilder::float64,
           Type.STRING, SchemaBuilder::string,
